+++ tests/binary_bilateral_test.py
@@ -6,6 +6,7 @@ import unittest
 import fixtures
 import numpy as np
 
+from lymph import models
 from lymph.modalities import Clinical
 
 
@@ -58,6 +59,33 @@ class BilateralInitTest(fixtures.BilateralModelMixin, unittest.TestCase):
             self.model.contra.modalities["foo"].specificity,
         )
 
+    def test_asymmetric_model(self):
+        """Check if different graphs work for the ipsi and contra side."""
+        ipsi_graph = fixtures.get_graph("medium")
+        contra_graph = fixtures.get_graph("small")
+
+        model = models.Bilateral(
+            graph_dict=ipsi_graph,
+            contralateral_kwargs={"graph_dict": contra_graph},
+        )
+
+        self.assertEqual(
+            list(model.ipsi.graph.nodes.keys()),
+            [key[1] for key in ipsi_graph.keys()],
+        )
+        self.assertEqual(
+            list(model.contra.graph.nodes.keys()),
+            [key[1] for key in contra_graph.keys()],
+        )
+        self.assertEqual(
+            len(model.ipsi.get_params()),
+            sum(len(val) for val in ipsi_graph.values()),
+        )
+        self.assertEqual(
+            len(model.contra.get_params()),
+            sum(len(val) for val in contra_graph.values()),
+        )
+
 
 class ModalityDelegationTestCase(fixtures.BilateralModelMixin, unittest.TestCase):
     """Make sure the modality is delegated from the ipsi side correctly."""
@@ -108,6 +136,82 @@ class ModalityDelegationTestCase(fixtures.BilateralModelMixin, unittest.TestCase
             self.model.contra.modalities["foo"].specificity,
         )
 
+    def test_diag_time_dists_delegation(self):
+        """Test that the diagnose time distributions are delegated."""
+        self.assertTrue(np.allclose(
+            list(self.model.diag_time_dists["early"].distribution),
+            list(self.model.ipsi.diag_time_dists["early"].distribution),
+        ))
+        self.assertTrue(np.allclose(
+            list(self.model.diag_time_dists["late"].get_params()),
+            list(self.model.ipsi.diag_time_dists["late"].get_params()),
+        ))
+        self.assertTrue(np.allclose(
+            list(self.model.diag_time_dists["early"].distribution),
+            list(self.model.contra.diag_time_dists["early"].distribution),
+        ))
+        self.assertTrue(np.allclose(
+            list(self.model.diag_time_dists["late"].get_params()),
+            list(self.model.contra.diag_time_dists["late"].get_params()),
+        ))
+
+
+class ParameterAssignmentTestCase(fixtures.BilateralModelMixin, unittest.TestCase):
+    """Test the parameter assignment."""
+
+    def setUp(self):
+        self.model_kwargs = {
+            "is_symmetric": {
+                "tumor_spread": False,
+                "lnl_spread": False,
+                "modalities": True,
+            }
+        }
+        super().setUp()
+
+    def test_get_params_as_args(self):
+        """Test that the parameters can be retrieved."""
+        ipsi_args = self.model.ipsi.get_params()
+        contra_args = self.model.contra.get_params()
+        self.assertEqual(len(ipsi_args), len(contra_args))
+
+    def test_get_params_as_dict(self):
+        """Test that the parameters can be retrieved."""
+        ipsi_dict = self.model.ipsi.get_params(as_dict=True)
+        contra_dict = self.model.contra.get_params(as_dict=True)
+        self.assertEqual(ipsi_dict.keys(), contra_dict.keys())
+
+    def test_assign_params_as_args(self):
+        """Test that the parameters can be assigned."""
+        ipsi_args = self.rng.uniform(size=len(self.model.ipsi.get_params()))
+        contra_args = self.rng.uniform(size=len(self.model.contra.get_params()))
+        none_args = [None] * len(ipsi_args)
+
+        # Assigning only the ipsi side
+        self.model.assign_params(*ipsi_args, *none_args)
+        self.assertTrue(np.allclose(ipsi_args, list(self.model.ipsi.get_params())))
+        self.assertEqual(
+            list(self.model.ipsi.diag_time_dists["late"].get_params())[0],
+            list(self.model.contra.diag_time_dists["late"].get_params())[0],
+        )
+
+        # Assigning only the contra side
+        self.model.assign_params(*none_args, *contra_args)
+        self.assertTrue(np.allclose(contra_args, list(self.model.contra.get_params())))
+        self.assertEqual(
+            list(self.model.ipsi.diag_time_dists["late"].get_params())[0],
+            list(self.model.contra.diag_time_dists["late"].get_params())[0],
+        )
+
+        # Assigning both sides
+        self.model.assign_params(*ipsi_args, *contra_args)
+        self.assertTrue(np.allclose(ipsi_args[:-1], list(self.model.ipsi.get_params())[:-1]))
+        self.assertTrue(np.allclose(contra_args, list(self.model.contra.get_params())))
+        self.assertEqual(
+            list(self.model.ipsi.diag_time_dists["late"].get_params())[0],
+            list(self.model.contra.diag_time_dists["late"].get_params())[0],
+        )
+
 
 class LikelihoodTestCase(fixtures.BilateralModelMixin, unittest.TestCase):
     """Check that the (log-)likelihood is computed correctly."""
@@ -135,10 +239,12 @@ class RiskTestCase(fixtures.BilateralModelMixin, unittest.TestCase):
         """Create a random diagnosis for each modality and LNL."""
         diagnoses = {}
 
-        for modality in self.model.modalities:
-            diagnoses[modality] = {}
-            for lnl in self.model.ipsi.graph.lnls.keys():
-                diagnoses[modality][lnl] = self.rng.choice([True, False, None])
+        for side in ["ipsi", "contra"]:
+            diagnoses[side] = {}
+            side_model = getattr(self.model, side)
+            lnl_names = side_model.graph.lnls.keys()
+            for modality in side_model.modalities:
+                diagnoses[side][modality] = fixtures.create_random_pattern(lnl_names)
 
         return diagnoses
 
@@ -155,3 +261,22 @@ class RiskTestCase(fixtures.BilateralModelMixin, unittest.TestCase):
         self.assertEqual(posterior.shape, (num_states, num_states))
         self.assertEqual(posterior.dtype, float)
         self.assertTrue(np.isclose(posterior.sum(), 1.))
+
+    def test_risk(self):
+        """Test that the risk is computed correctly."""
+        random_parameters = self.create_random_params()
+        random_diagnoses = self.create_random_diagnoses()
+        random_pattern = {
+            "ipsi": fixtures.create_random_pattern(self.model.ipsi.graph.lnls.keys()),
+            "contra": fixtures.create_random_pattern(self.model.contra.graph.lnls.keys()),
+        }
+        random_t_stage = self.rng.choice(["early", "late"])
+
+        risk = self.model.risk(
+            involvement=random_pattern,
+            given_param_kwargs=random_parameters,
+            given_diagnoses=random_diagnoses,
+            t_stage=random_t_stage,
+        )
+        self.assertLessEqual(risk, 1.)
+        self.assertGreaterEqual(risk, 0.)