+++ tests/trinary_unilateral_test.py
@@ -1,21 +1,32 @@
 """Test the trinary unilateral system."""
 import unittest
 
+import fixtures
 import numpy as np
 import pandas as pd
 
-from tests.fixtures import TrinaryFixtureMixin
+from lymph.graph import LymphNodeLevel
 
 
-class TrinaryInitTestCase(TrinaryFixtureMixin, unittest.TestCase):
+class TrinaryInitTestCase(fixtures.TrinaryFixtureMixin, unittest.TestCase):
     """Testing the basic initialization of a trinary model."""
 
     def test_is_trinary(self) -> None:
         """Test if the model is trinary."""
         self.assertTrue(self.model.is_trinary)
 
+    def test_lnls(self):
+        """Test they are all trinary lymph node levels."""
+        model_allowed_states = self.model.graph.allowed_states
+        self.assertEqual(len(model_allowed_states), 3)
 
-class TrinaryTransitionMatrixTestCase(TrinaryFixtureMixin, unittest.TestCase):
+        for lnl in self.model.graph.lnls.values():
+            self.assertIsInstance(lnl, LymphNodeLevel)
+            self.assertTrue(lnl.is_trinary)
+            self.assertEqual(lnl.allowed_states, model_allowed_states)
+
+
+class TrinaryTransitionMatrixTestCase(fixtures.TrinaryFixtureMixin, unittest.TestCase):
     """Test the transition matrix of a trinary model."""
 
     def setUp(self):
@@ -48,7 +59,7 @@ class TrinaryTransitionMatrixTestCase(TrinaryFixtureMixin, unittest.TestCase):
         self.assertTrue(np.allclose(row_sums, 1.0))
 
 
-class TrinaryObservationMatrixTestCase(TrinaryFixtureMixin, unittest.TestCase):
+class TrinaryObservationMatrixTestCase(fixtures.TrinaryFixtureMixin, unittest.TestCase):
     """Test the observation matrix of a trinary model."""
 
     def setUp(self):
@@ -68,7 +79,7 @@ class TrinaryObservationMatrixTestCase(TrinaryFixtureMixin, unittest.TestCase):
         self.assertTrue(np.allclose(row_sums, 1.0))
 
 
-class TrinaryDiagnoseMatricesTestCase(TrinaryFixtureMixin, unittest.TestCase):
+class TrinaryDiagnoseMatricesTestCase(fixtures.TrinaryFixtureMixin, unittest.TestCase):
     """Test the diagnose matrix of a trinary model."""
 
     def setUp(self):
@@ -87,3 +98,64 @@ class TrinaryDiagnoseMatricesTestCase(TrinaryFixtureMixin, unittest.TestCase):
             num_patients = (self.model.patient_data["_model", "#", "t_stage"] == t_stage).sum()
             diagnose_matrix = self.model.diagnose_matrices[t_stage]
             self.assertEqual(diagnose_matrix.shape, (3 ** num_lnls, num_patients))
+
+
+class TrinaryLikelihoodTestCase(fixtures.TrinaryFixtureMixin, unittest.TestCase):
+    """Test the likelihood of a trinary model."""
+
+    def setUp(self):
+        """Load patient data."""
+        super().setUp()
+        self.model.modalities = fixtures.MODALITIES
+        self.init_diag_time_dists(early="frozen", late="parametric")
+        self.model.assign_params(**self.create_random_params())
+        self.load_patient_data(filename="2021-usz-oropharynx.csv")
+
+    def test_log_likelihood_smaller_zero(self):
+        """Make sure the log-likelihood is smaller than zero."""
+        likelihood = self.model.likelihood(log=True, mode="HMM")
+        self.assertLess(likelihood, 0.)
+
+    def test_likelihood_invalid_params_isinf(self):
+        """Make sure the likelihood is `-np.inf` for invalid parameters."""
+        random_params = self.create_random_params()
+        for name in random_params:
+            random_params[name] += 1.
+        likelihood = self.model.likelihood(
+            given_param_kwargs=random_params,
+            log=True,
+            mode="HMM",
+        )
+        self.assertEqual(likelihood, -np.inf)
+
+
+class TrinaryRiskTestCase(fixtures.TrinaryFixtureMixin, unittest.TestCase):
+    """Test the risk of a trinary model."""
+
+    def setUp(self):
+        """Load patient data."""
+        super().setUp()
+        self.model.modalities = fixtures.MODALITIES
+        self.init_diag_time_dists(early="frozen", late="parametric")
+        self.load_patient_data(filename="2021-usz-oropharynx.csv")
+
+    def create_random_diagnoses(self):
+        """Create a random diagnosis for each modality and LNL."""
+        lnl_names = list(self.model.graph.lnls.keys())
+        diagnoses = {}
+
+        for modality in self.model.modalities:
+            diagnoses[modality] = fixtures.create_random_pattern(lnl_names)
+
+        return diagnoses
+
+    def test_risk_is_probability(self):
+        """Make sure the risk is a probability."""
+        risk = self.model.risk(
+            involvement=fixtures.create_random_pattern(lnls=list(self.model.graph.lnls.keys())),
+            given_diagnoses=self.create_random_diagnoses(),
+            given_param_kwargs=self.create_random_params(),
+            t_stage=self.rng.choice(["early", "late"]),
+        )
+        self.assertGreaterEqual(risk, 0.)
+        self.assertLessEqual(risk, 1.)